Re-raise the original exception when limits are reached
Previously when retrying continued to fail a LimitReached exception was
raised. This unfrotunately meant the original exception was lost and
code needed to be updated to handle a retrace specific exception.

This change now causes the decorator to re-raise the last exception that
was caught when the retry limit is reached.

Closes #4